feat(timeline): Allow adding frames from asset library
Implement drag-and-drop functionality from the asset library to the timeline. This allows users to select one or more assets and drop them into the timeline to create new frames. The new frames will be populated with the selected asset's image and file. This feature streamlines the process of adding existing assets to the video storyboard.