feat(asset-library): Enable dragging sketches to asset library
This change modifies the `AssetLibraryPanel` to accept dropped sketches. When a sketch is dropped, it will be added as a new asset to the library. This allows users to easily convert existing sketches into reusable assets.

Additionally, the `SketchCard` component has been updated to be draggable, emitting drag events that can be captured by the `AssetLibraryPanel`.